intelgraph-mvp/api/pyproject.toml
[project]
name = "intelgraph-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings",
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-jose[cryptography]",
    "neo4j",
    "python-multipart",
]

[project.optional-dependencies]
dev = ["pytest", "httpx", "pytest-asyncio"]
